move e2e helper functions out of internal package
