+++ test/e2e/gardener/project/project.go
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package internal
+package project
 
 import (
 	"time"
