+++ test/e2e/operator/garden/garden.go
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package internal
+package garden
 
 import (
 	"fmt"
